# src/be/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DpSolution API",
        "description": "Invoice management sistem API endpoints",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/v1/tokens": {
            "get": {
                "tags": [
                    "Tokens"
                ],
                "summary": "Get list of tokens",
                "description": "Returns list of tokens",
                "operationId": "getTokensList",
                "parameters": [
                    {
                        "name": "include_tags",
                        "in": "query",
                        "description": "Include tags in the response",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "tokenAddress[lk]",
                        "in": "query",
                        "description": "Search query [eq,lt,lte,gt,gte,ne,lk,ilk,nlk,inlk,bt,nbt,in,nin,json]",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-02-15T23:21:32+04:00"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "api/v1/tokens"
                                        },
                                        "method": {
                                            "type": "string",
                                            "example": "GET"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "result": {
                                            "$ref": "#/components/schemas/TokenCollection"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tokens"
                ],
                "summary": "Store new token",
                "description": "Returns token data",
                "operationId": "storeToken",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTokenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-02-15T23:21:32+04:00"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "api/v1/tokens"
                                        },
                                        "method": {
                                            "type": "string",
                                            "example": "POST"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "result": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Token successfully created"
                                                },
                                                "data": {
                                                    "$ref": "#/components/schemas/TokenResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-02-15T23:21:32+04:00"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "api/v1/tokens"
                                        },
                                        "method": {
                                            "type": "string",
                                            "example": "POST"
                                        },
                                        "error": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "Address information is not recorded."
                                                }
                                            }
                                        },
                                        "result": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreTokenRequest": {
                "title": "Store Token Request",
                "description": "Store Token request body data",
                "required": [
                    "tokenAddress",
                    "tokenName",
                    "tokenSymbol",
                    "tokenTotalSupply"
                ],
                "properties": {
                    "tokenAddress": {
                        "description": "The unique address of the token",
                        "type": "string",
                        "example": "0x1234567890abcdef1234567890abcdef12345678"
                    },
                    "tokenName": {
                        "description": "The name of the token",
                        "type": "string",
                        "example": "MyToken"
                    },
                    "tokenSymbol": {
                        "description": "The symbol of the token",
                        "type": "string",
                        "example": "MTK"
                    },
                    "tokenTotalSupply": {
                        "description": "Total supply of the token (decimal value supported)",
                        "type": "number",
                        "format": "decimal",
                        "example": 1000000.5
                    }
                },
                "type": "object"
            },
...
            "TokenCollection": {
                "title": "Token Collection",
                "description": "Token collection",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Token successfully full list."
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TokenResource"
                        }
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "from": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 15
                            },
                            "to": {
                                "type": "integer",
                                "example": 15
                            },
                            "total": {
                                "type": "integer",
                                "example": 50
                            }
                        },
                        "type": "object"
                    },
                    "links": {
                        "properties": {
                            "self": {
                                "type": "string",
                                "example": "http://example.com/api/tokens?page=1"
                            },
                            "first": {
                                "type": "string",
                                "example": "http://example.com/api/tokens?page=1"
                            },
                            "last": {
                                "type": "string",
                                "example": "http://example.com/api/tokens?page=4"
                            },
                            "prev": {
                                "type": "string",
                                "nullable": true
                            },
                            "next": {
                                "type": "string",
                                "example": "http://example.com/api/tokens?page=2"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "TokenResource": {
                "title": "Token Resource",
                "description": "Token resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tokenAddress": {
                        "description": "The unique blockchain address of the token",
                        "type": "string",
                        "example": "0x1234567890abcdef1234567890abcdef12345678"
                    },
                    "tokenName": {
                        "description": "The name of the token",
                        "type": "string",
                        "example": "MyToken"
                    },
                    "tokenSymbol": {
                        "description": "The symbol of the token",
                        "type": "string",
                        "example": "MTK"
                    },
                    "tokenTotalSupply": {
                        "description": "Total supply of the token (decimal value supported)",
                        "type": "number",
                        "format": "decimal",
                        "example": 1000000.5
                    },
                    "createdAt": {
                        "description": "Timestamp when the token was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01 00:00:00"
                    },
                    "updatedAt": {
                        "description": "Timestamp when the token was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-02 12:30:00"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}